backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
shapely==2.0.4
pvlib==0.10.5
numpy
pandas
scikit-learn
xgboost
lightgbm
joblib
requests==2.32.3
httpx==0.27.0
redis==5.0.5
celery==5.3.6
reportlab==4.2.0
Pillow==10.3.0
pyproj==3.6.1
opencv-python==4.11.0.86
overpy==0.7
aiohttp==3.9.5
python-multipart==0.0.9
supabase>=2.0.0
PyJWT>=2.0.0
pytest==8.2.1
pytest-asyncio==0.23.7
